% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module-auth.R
\name{auth_labels}
\alias{auth_labels}
\title{Labels for authentication module}
\usage{
auth_labels(please_authenticate = "Please authenticate",
  username = "Username:", password = "Password:", login = "Login",
  invalid_usr_pwd = "Username or password are incorrect",
  user_expired = "Your account has expired")
}
\arguments{
\item{please_authenticate}{Text to be displayed as a header.}

\item{username}{Label for username input field.}

\item{password}{Label for password input field.}

\item{login}{Text displayed on the button.}

\item{invalid_usr_pwd}{Error message displayed if authentication is unsuccesful.}

\item{user_expired}{Error message displayed if user's account has expired.}
}
\description{
Labels for authentication module
}
