% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_credentials.R
\name{check-credentials}
\alias{check-credentials}
\alias{check_credentials_df}
\title{Check credentials}
\usage{
check_credentials_df(user, password, credentials_df)
}
\arguments{
\item{user}{User to authenticate.}

\item{password}{Password of user to authenticate.}

\item{credentials_df}{A \code{data.frame} containing credentials,
must contain columns \code{user} and \code{password}.}
}
\value{
A \code{list} with 2 slots :
 \itemize{
  \item \strong{result} : logical, result of authentication.
  \item \strong{user_info} : the line in \code{credentials_df} corresponding to the user.
 }
}
\description{
Check credentials
}
\examples{
# data.frame with credentials info
credentials <- data.frame(
  user = c("fanny", "victor"),
  password = c("azerty", "12345"),
  stringsAsFactors = FALSE
)

# check a user
check_credentials_df("fanny", "azerty", credentials)
check_credentials_df("fanny", "azert", credentials)
check_credentials_df("fannyyy", "azerty", credentials)

check_credentials_p <- purrr::partial(
  check_credentials_df,
  credentials_df = credentials # set default df to use
)
check_credentials_p("fanny", "azerty")
}
