% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials-db.R
\name{create_db}
\alias{create_db}
\title{Create credentials database}
\usage{
create_db(credentials_data, sqlite_path, passphrase = NULL)
}
\arguments{
\item{credentials_data}{A \code{data.frame}}

\item{sqlite_path}{Path to the SQLite database.}

\item{passphrase}{A password to protect the data inside the database.}
}
\description{
Create credentials database
}
