% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials-db.R
\name{create_sql_tables}
\alias{create_sql_tables}
\title{Create credentials tables in PostgreSQL database}
\usage{
create_sql_tables(conn, credentials_data, passphrase = NULL)
}
\arguments{
\item{conn}{SQL database connection}

\item{credentials_data}{credentials data.frame}

\item{passphrase}{option passphrase for database encryption}
}
\value{
A data.frame
}
\description{
Create credentials tables in PostgreSQL database
}
\examples{
\dontrun{

# PostgreSQL Connection
conn <- DBI::dbConnect(RPostgres::Postgres(),
                       host   = "host",
                       dbname = "db_name",
                       user = "user",
                       password = "pw",
                       port = 5432)

# Credentials data.frame
credentials <- data.frame(user = c("shiny", "admin"),
                          password = c("shiny", "admin"),
                          start = c(NA, NA),
                          expire = c(NA, NA),
                          admin = c(FALSE, TRUE),
                          stringsAsFactors = FALSE)

# Write SQL tables
create_sql_tables(conn,
                  credentials_data = credentials,
                  passphrase = NULL)

}
}
