% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_access_keys_2}
\alias{custom_access_keys_2}
\title{custom_access_keys_2}
\usage{
custom_access_keys_2(
  name_of_secret,
  path_to_keys_db = "../../base-data/database/keys_database.sqlite",
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{name_of_secret}{Name of the secret to be loaded.}

\item{path_to_keys_db}{Path to keys_database.sqlite (optional).}

\item{path_to_user_db}{Path to shiny_users.sqlite (optional).}
}
\value{
Decrypted secret for the requested_data.
}
\description{
Loads secret for given name from the keys database and returns the
decrypted secret. 
When triggered interactively, asks for password of produkt user to decrypt the secret.
On the server, the provided password of the signed in user is used.
}
