% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_add_secret}
\alias{custom_add_secret}
\title{custom_add_secret}
\usage{
custom_add_secret(
  name,
  new_secret,
  description = "",
  path_to_keys_db = "../../base-data/database/keys_database.sqlite",
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{name}{Name of the new secret}

\item{new_secret}{Secret to be stored}

\item{description}{Description}

\item{path_to_keys_db}{Path to keys_database.sqlite}

\item{path_to_user_db}{Path to shiny_users.sqlite}
}
\description{
Encrypts new secret with master_key and stores it in the keys_database
}
