% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_decrypt_data_2}
\alias{custom_decrypt_data_2}
\title{custom_decrypt_data_2}
\usage{
custom_decrypt_data_2(
  encrypted_df,
  name_of_secret,
  path_to_keys_db = "../../base-data/database/keys_database.sqlite",
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{encrypted_df}{Encrypted data frames.}

\item{name_of_secret}{The name of the secret that decrypts the data.}

\item{path_to_keys_db}{Path to keys_database.sqlite (optional).}

\item{path_to_user_db}{Path to shiny_users.sqlite (optional).}
}
\value{
The decrypted dataframe
}
\description{
Decrypts a dataframe without having to provide the secret.
}
