% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_delete_secret}
\alias{custom_delete_secret}
\title{custom_delete_secret}
\usage{
custom_delete_secret(
  secret,
  path_to_keys_db = "../../base-data/database/keys_database.sqlite"
)
}
\arguments{
\item{path_to_keys_db}{Path to keys_database.sqlite}

\item{name}{Name of the secret}
}
\description{
Deletes a secret from the keys_database
}
