% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_encrypt_data}
\alias{custom_encrypt_data}
\title{custom_encrypt_data}
\usage{
custom_encrypt_data(
  data_df,
  name_of_secret,
  path_to_keys_db = "../../base-data/database/keys_database.sqlite",
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{data_df}{The data frame to be encrypted.}

\item{name_of_secret}{The name of the secret that decrypts the data.}

\item{path_to_keys_db}{Path to keys_database.sqlite (optional).}

\item{path_to_user_db}{Path to shiny_users.sqlite (optional).}
}
\value{
The encrypted dataframe .
}
\description{
Encrypts a dataframe completely without having to provide the secret.
}
