% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_encrypt_db}
\alias{custom_encrypt_db}
\title{custom_encrypt_db}
\usage{
custom_encrypt_db(df, name_of_secret, columns_to_encrypt)
}
\arguments{
\item{df}{The dataframe to be encrypted.}

\item{name_of_secret}{The name of the secret that encrypts the data.}

\item{columns_to_encrypt}{The columns that need to be encrpyted.}
}
\description{
This function is used to encrypt a dataframe that is saved within a sqlite database.
You are able to specifiy which columns you want to encrypt.
}
