% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_filter_teamleads}
\alias{custom_filter_teamleads}
\title{custom_filter_teamlead}
\usage{
custom_filter_teamleads(
  sales_teams,
  employee,
  user_permission,
  database_needed
)
}
\arguments{
\item{sales_teams}{dataframe containing information about all sales employees}

\item{employee}{sales employee/logged in user name}

\item{user_permission}{permission level of the user}

\item{datebase_needed}{Logical value (TRUE or FALSE) indicating if the database format is needed.}
}
\value{
return name of vector of names of Sales Mitarbeiter in Team
}
\description{
Filters sales_teams dataframe to extract only Sales MA in 
specific team of Teamlead (employee).
}
