% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_interactive}
\alias{custom_interactive}
\title{custom_interactive}
\usage{
custom_interactive()
}
\value{
Boolean value indicating if a function is called from a shiny app or console
}
\description{
This function checks if a function is called from a shiny app or from the console.
It works just like interactive(), but returns FALSE if running from a shiny app.
}
