% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_load_data_in_module}
\alias{custom_load_data_in_module}
\title{custom_load_shiny_module_data}
\usage{
custom_load_data_in_module(data_file, name_of_secret)
}
\arguments{
\item{data_file}{This can be an Data Frame for example TestData or the path (string) to the data file}

\item{name_of_secret}{Only necessary if the data is encrypted: The name of the secret in the keys database that decrypts the data}
}
\value{
Returns the loaded data as a dataframe.
}
\description{
custom_load_shiny_module_data
}
