% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_permission}
\alias{custom_permission}
\title{custom_permission}
\usage{
custom_permission(
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{path_to_user_db}{= "../../base-data/database/shiny_users.sqlite" (optional)}
}
\value{
The Level of Rights the user has
}
\description{
Returns an integer telling you hat level of permission the user has.
This function does not need any parameters.
}
