% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_permission_level}
\alias{custom_permission_level}
\title{custom_permission_level}
\usage{
custom_permission_level(
  path_to_user_db = "../../base-data/database/shiny_users.sqlite"
)
}
\arguments{
\item{path_to_user_db}{The path to the user database. Default is "../../base-data/database/shiny_users.sqlite". (optional)}
}
\value{
An integer representing the user's permission level.
}
\description{
Returns an integer telling you hat level of permission the user has.
This function does not need any parameters.
}
