% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_retrieve_user_role_v2}
\alias{custom_retrieve_user_role_v2}
\title{custom_retrieve_user_role_v2}
\usage{
custom_retrieve_user_role_v2(
  user_name,
  path_to_user_db = "../../base-data/database/shiny_users.sqlite",
  preset_key = NA_character_
)
}
\arguments{
\item{path_to_user_db}{Path to the SQLite database containing user credentials.}

\item{preset_key}{The key to decrypt the data. Default is NA_character_. Please Remove the Key immediately from the environment after your authentication. (optional)}
}
\value{
Returns the role name of the user.
}
\description{
Retrieves the role of a user from the database.
}
