% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_show_secrets}
\alias{custom_show_secrets}
\title{custom_show_secrets}
\usage{
custom_show_secrets(
  path_to_keys_db = "../../base-data/database/keys_database.sqlite"
)
}
\arguments{
\item{path_to_keys_db}{path to keys_database.sqlite}
}
\value{
data data frame with names and descriptions
}
\description{
encryps secret using master key and stores it in the local database
}
