% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_username}
\alias{custom_username}
\title{custom_username}
\usage{
custom_username(preset_key = NA_character_)
}
\arguments{
\item{preset_key}{The key to decrypt the data. Default is NA_character_. Please Remove the Key immediately from the environment after your authentication. (optional)}
}
\value{
The username
}
\description{
This function is used to retrieve the username of the user.
}
