% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_pwd}
\alias{generate_pwd}
\title{Simple password generation}
\usage{
generate_pwd(n = 1)
}
\arguments{
\item{n}{Number of password(s)}
}
\value{
a \code{character}
}
\description{
Simple password generation
}
\examples{
generate_pwd()

generate_pwd(3)
}
