% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_db.R
\name{initialize_db}
\alias{initialize_db}
\title{Initialize credentials database}
\usage{
initialize_db(credentials_data, sqlite_path, passphrase = NULL)
}
\arguments{
\item{credentials_data}{A \code{data.frame}}

\item{sqlite_path}{Path to the SQLite database.}

\item{passphrase}{A password to protect the data inside the database.}
}
\description{
Initialize credentials database
}
