% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage-auth.R
\name{manage-auth}
\alias{manage-auth}
\alias{manage_auth_app}
\alias{manage_auth_server}
\title{Manage authentication in a Shiny application}
\usage{
manage_auth_app(ui, ...)

manage_auth_server(session, check_credentials)
}
\arguments{
\item{ui}{UI of the application.}

\item{...}{Arguments passed to \code{\link{auth_ui}}.}

\item{session}{Shiny session}

\item{check_credentials}{Function passed to \code{\link{auth_server}}.}
}
\description{
Manage authentication in a Shiny application
}
