% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials-db.R
\name{read_sql_decrypt}
\alias{read_sql_decrypt}
\title{Read decrypted table from a Postgres database}
\usage{
read_sql_decrypt(conn, name = "credentials", passphrase = NULL)
}
\arguments{
\item{conn}{SQL database connection}

\item{name}{SQL table name}

\item{passphrase}{option passphrase for database encryption}
}
\value{
A data.frame
}
\description{
Read decrypted table from a Postgres database
}
