% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{use_language}
\alias{use_language}
\title{Use shinymanager labels}
\usage{
use_language(lan = "en")
}
\arguments{
\item{lan}{Language to use for labels, supported values are : "en", "fr", "pt-BR", "es", "de", "pl", "ja", "el", "id", "zh-CN", "no", "it".}
}
\value{
A language object
}
\description{
See all labels registered with \code{get_labels()},
 then set custom text with \code{set_labels()}.
}
\examples{

use_language(lan = "fr")

}
