% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials-db.R
\name{write_sql_encrypt}
\alias{write_sql_encrypt}
\title{Write encrypted table to a Postgres database}
\usage{
write_sql_encrypt(conn, value, name = "credentials", passphrase = NULL)
}
\arguments{
\item{conn}{SQL database connection}

\item{value}{data.frame to write as SQL table}

\item{name}{SQL table name}

\item{passphrase}{option passphrase for database encryption}
}
\value{
A data.frame
}
\description{
Write encrypted table to a Postgres database
}
