% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_access_keys}
\alias{custom_access_keys}
\title{custom_access_keys}
\usage{
custom_access_keys(requested_data)
}
\arguments{
\item{requested_data}{Name of the secret you want to access}
}
\value{
Decrypted secret for the requested_data
}
\description{
Decrypts master_key using users password (entered through shinymanager ui) from the shiny_users.sqlite database
Afterwards decrypts requested data with the decrypted master_key from the keys_database.sqlite
Function is deprecated, please use custom_access_keys_2.
}
