% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_access_keys}
\alias{custom_access_keys}
\title{custom_access_keys}
\usage{
custom_access_keys(requested_data)
}
\arguments{
\item{requested_data}{name of the key you want to access}

\item{encrypted_master_key}{master_key decrypted with users password}

\item{key}{users password}
}
\value{
decryption key for a data set
}
\description{
decryps master key using users password and then uses master key to decrypt the requested data
}
