% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_add_secret}
\alias{custom_add_secret}
\title{custom_add_secret}
\usage{
custom_add_secret(
  name,
  new_secret,
  description = "",
  path_to_keys_db = "../../base-data/database/keys_database.sqlite",
  path_to_user_db = "../../base-data/database/shiny_users"
)
}
\arguments{
\item{name}{name of the new secret}

\item{new_secret}{new secret}

\item{description}{description}

\item{path_to_keys_db}{path to keys_database.sqlite}

\item{path_to_user_db}{path to shiny_users.sqlite}
}
\description{
encrypts and stores new secret in keys_database
}
