% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_add_secret}
\alias{custom_add_secret}
\title{custom_add_secret}
\usage{
custom_add_secret(name, new_secret, description)
}
\arguments{
\item{name}{name of the new secret}

\item{new_secret}{new secret}

\item{description}{description}
}
\description{
encryps and stores new secret in keys_database
}
