% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_decrypt_db}
\alias{custom_decrypt_db}
\title{custom_decrypt_db}
\usage{
custom_decrypt_db(df, name_of_secret, columns_to_encrypt)
}
\arguments{
\item{df}{The dataframe to be decrypted.}

\item{name_of_secret}{The name of the secret that decrypts the data.}

\item{columns_to_encrypt}{The columns that need to be decrpyted.}
}
\description{
This function is used to decrypt a dataframe that is saved within a sqlite database.
You are able to specifiy which columns you want to decrypt.
}
