% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_delete_secret}
\alias{custom_delete_secret}
\title{custom_delete_secret}
\usage{
custom_delete_secret(secret)
}
\arguments{
\item{name}{name of the new secret}
}
\description{
returns names of stored secrets and descriptions
}
