% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_delete_secret}
\alias{custom_delete_secret}
\title{custom_delete_secret}
\usage{
custom_delete_secret(
  name_of_secret,
  path_to_keys_db = "../../base-data/database/keys_database.sqlite"
)
}
\arguments{
\item{name_of_secret}{Name of the secret to be deleted.}

\item{path_to_keys_db}{Path to keys_database.sqlite (optional).}
}
\value{
Message indicating the secret has been deleted.
}
\description{
Should only be triggered interactively.
Deletes a secret from the keys_database
}
