% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_get_user_scope}
\alias{custom_get_user_scope}
\title{Get User Scope from PostgreSQL}
\usage{
custom_get_user_scope(con, user_name, user_role)
}
\arguments{
\item{con}{A valid PostgreSQL database connection (e.g., from `DBI::dbConnect`).}

\item{user_name}{the name of the user from the login}

\item{user_role}{the user role - can be retrieved via `custom_retrieve_user_role()` but should be done before the function call}
}
\value{
A named list with vectors of service user IDs for the following services:
\describe{
  \item{crm}{User IDs for the CRM system.}
  \item{billomat_templates}{User IDs for Billomat templates.}
  \item{gsheet_videocalls}{User IDs for Google Sheets Video Calls.}
  \item{sipgate}{User IDs for Sipgate.}
  \item{msgraph}{User IDs for Microsoft Graph.}
}
}
\description{
Retrieves the user-specific service scopes from the database based on the authenticated user name
and their role. The function queries the `get_service_users_in_scope` database function and
returns a list of user IDs for each connected service.
}
\details{
The user role is automatically retrieved via `custom_retrieve_user_role()`. If the user role
falls under certain privileged roles, a placeholder user may be used in the query inside the database.
}
