% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_retrieve_credentials}
\alias{custom_retrieve_credentials}
\title{custom_retrieve_credentials}
\usage{
custom_retrieve_credentials(username = TRUE, password = TRUE)
}
\arguments{
\item{username}{Boolean indicating if the user_name should be retrieved. Default is TRUE (optional).}

\item{password}{Boolean indicating if the password should be retrieved. Default is TRUE (optional).}
}
\value{
List which contains the user_name and password
}
\description{
This function is used to retrieve the credentials of the user.
}
