% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_show_secrets}
\alias{custom_show_secrets}
\title{custom_show_secrets}
\usage{
custom_show_secrets(
  path_to_keys_db = "../../base-data/database/keys_database.sqlite"
)
}
\arguments{
\item{path_to_keys_db}{Path to keys_database.sqlite (optional).}
}
\value{
Dataframe with names and descriptions of stored secrets.
}
\description{
Should only be triggered interactively.
Retrieves all stored names and descriptions from the keys_database.
}
