% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_show_secrets}
\alias{custom_show_secrets}
\title{custom_show_secrets}
\usage{
custom_show_secrets()
}
\value{
data data frame with names and descriptions
}
\description{
encryps secret using master key and stores it in the local database
}
