% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_functions.R
\name{custom_show_warnings}
\alias{custom_show_warnings}
\title{custom_show_warnings}
\usage{
custom_show_warnings()
}
\description{
This function shows the warning of any custom function.
If a shiny app is started, the warning pops up on the dashboard.
If a function is called locally, the warning is printed on the console.
}
