% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage-auth.R
\name{manage-auth}
\alias{manage-auth}
\alias{manage_auth_app}
\alias{manage_auth_server}
\title{Manage authentication in a Shiny application}
\usage{
manage_auth_app(ui, ..., enable_admin = FALSE, head_auth = NULL)

manage_auth_server(check_credentials,
  session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{ui}{UI of the application.}

\item{...}{Arguments passed to \code{\link{auth_ui}}.}

\item{enable_admin}{Enable or not access to admin mode, note that
admin mode is only available using a SQLite backend for credentials.}

\item{head_auth}{Tag or list of tags to use in the \code{<head>}
of the authentication page (for custom CSS for example).}

\item{check_credentials}{Function passed to \code{\link{auth_server}}.}

\item{session}{Shiny session.}
}
\description{
Manage authentication in a Shiny application
}
